requirements.txt
aiogram==3.18.0
PyMuPDF==1.25.3
python-telegram-bot==21.11.1
python-dotenv==1.0.0
mysql-connector-python
python-dotenv